backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.7.1
celery==5.4.0
SQLAlchemy==2.0.36
httpx==0.28.1
beautifulsoup4==4.12.3
networkx==3.4.2
requests==2.32.3
werkzeug==3.1.3
python-dotenv==1.0.1
redis==5.2.1
lxml==5.3.0
PyPDF2>=3.0.0
exifread>=3.0.0
Pillow>=10.0.0
imagehash>=4.3.0
fpdf2>=2.7.0
